requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
alsa-lib=1.2.11=hd590300_1
attr=2.5.1=h166bdaf_1
barrnap=0.9=hdfd78af_4
bedtools=2.31.1=hf5e1c6e_1
biopython=1.83=py312h98912ed_0
bzip2=1.0.8=hd590300_5
ca-certificates=2024.2.2=hbcca054_0
cairo=1.18.0=h3faef2a_0
colorama=0.4.6=pyhd8ed1ab_0
dbus=1.13.6=h5008d03_3
ete3=3.1.3=pyh1d129d4_0
expat=2.6.2=h59595ed_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=h77eed37_1
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.12.1=h267a509_2
gawk=5.3.0=ha916aea_0
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h0b41bf4_3
glib=2.80.0=hf2295e7_1
glib-tools=2.80.0=hde27a5a_1
gmp=6.3.0=h59595ed_1
graphite2=1.3.13=h59595ed_1003
gst-plugins-base=1.22.9=hfa15dee_1
gstreamer=1.22.9=h98fc4e7_1
harfbuzz=8.3.0=h3d44ed6_0
hmmer=3.4=hdbdd923_1
icu=73.2=h59595ed_0
iqtree=2.3.0=h21ec9f0_0
keyutils=1.6.1=h166bdaf_0
krb5=1.21.2=h659d440_0
lame=3.100=h166bdaf_1003
ld_impl_linux-64=2.40=h41732ed_0
lerc=4.0.0=h27087fc_0
libblas=3.9.0=21_linux64_openblas
libcap=2.69=h0f662aa_0
libcblas=3.9.0=21_linux64_openblas
libclang-cpp15=15.0.7=default_h127d8a8_5
libclang13=18.1.2=default_h5d6823c_1
libcups=2.3.3=h4637d8d_4
libdeflate=1.20=hd590300_0
libedit=3.1.20191231=he28a2e2_2
libevent=2.1.12=hf998b51_1
libexpat=2.6.2=h59595ed_0
libffi=3.4.2=h7f98852_5
libflac=1.4.3=h59595ed_0
libgcc-ng=13.2.0=h807b86a_5
libgcrypt=1.10.3=hd590300_0
libgfortran-ng=13.2.0=h69a702a_5
libgfortran5=13.2.0=ha4646dd_5
libglib=2.80.0=hf2295e7_1
libgomp=13.2.0=h807b86a_5
libgpg-error=1.48=h71f35ed_0
libiconv=1.17=hd590300_2
libjpeg-turbo=3.0.0=hd590300_1
liblapack=3.9.0=21_linux64_openblas
libllvm15=15.0.7=hb3ce162_4
libllvm18=18.1.2=h2448989_0
libnsl=2.0.1=hd590300_0
libogg=1.3.4=h7f98852_1
libopenblas=0.3.26=pthreads_h413a1c8_0
libopus=1.3.1=h7f98852_1
libpng=1.6.43=h2797004_0
libpq=16.2=h33b98f1_1
libsndfile=1.2.2=hc60ed4a_1
libsqlite=3.45.2=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_5
libsystemd0=255=h3516f8a_1
libtiff=4.6.0=h1dd3fc0_3
libuuid=2.38.1=h0b41bf4_0
libvorbis=1.3.7=h9c3ff4c_0
libwebp=1.3.2=h658648e_1
libwebp-base=1.3.2=hd590300_0
libxcb=1.15=h0b41bf4_0
libxcrypt=4.4.36=hd590300_1
libxkbcommon=1.7.0=h662e7e4_0
libxml2=2.12.6=h232c23b_1
libxslt=1.1.39=h76b75d6_0
libzlib=1.2.13=hd590300_5
lxml=5.1.0=py312h37b5203_0
lz4-c=1.9.4=hcb278e6_0
mafft=7.525=h031d066_1
mpfr=4.2.1=h9458935_0
mpg123=1.32.4=h59595ed_0
mysql-common=8.3.0=hf1915f5_4
mysql-libs=8.3.0=hca2cd23_4
ncurses=6.4.20240210=h59595ed_0
nspr=4.35=h27087fc_0
nss=3.98=h1d7d5a4_0
numpy=1.26.4=py312heda63a1_0
openssl=3.2.1=hd590300_1
packaging=24.0=pyhd8ed1ab_0
pcre2=10.43=hcad00b1_0
perl=5.32.1=7_hd590300_perl5
pip=24.0=pyhd8ed1ab_0
pixman=0.43.2=h59595ed_0
ply=3.11=py_1
pthread-stubs=0.4=h36c2ea0_1001
pulseaudio-client=17.0=hb77b528_0
pyqt=5.15.9=py312h949fe66_5
pyqt5-sip=12.12.2=py312h30efb56_5
python=3.12.2=hab00c5b_0_cpython
python_abi=3.12=4_cp312
qt=5.15.8=hf11cfaa_0
qt-main=5.15.8=h112747c_20
qt-webengine=5.15.8=h3e791b3_6
readline=8.2=h8228510_1
scipy=1.12.0=py312heda63a1_2
setuptools=69.2.0=pyhd8ed1ab_0
sip=6.7.12=py312h30efb56_0
six=1.16.0=pyh6c4a22f_0
tk=8.6.13=noxft_h4845f30_101
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
tqdm=4.66.2=pyhd8ed1ab_0
tzdata=2024a=h0c530f3_0
wheel=0.43.0=pyhd8ed1ab_1
xcb-util=0.4.0=hd590300_1
xcb-util-image=0.4.0=h8ee46fc_1
xcb-util-keysyms=0.4.0=h8ee46fc_1
xcb-util-renderutil=0.3.9=hd590300_1
xcb-util-wm=0.4.1=h8ee46fc_1
xkeyboard-config=2.41=hd590300_0
xorg-compositeproto=0.4.2=h7f98852_1001
xorg-damageproto=1.2.1=h7f98852_1002
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.7=h8ee46fc_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxcomposite=0.4.6=h0b41bf4_1
xorg-libxdamage=1.1.5=h7f98852_1
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-libxrandr=1.5.2=h7f98852_1
xorg-libxrender=0.9.11=hd590300_0
xorg-libxtst=1.2.3=h7f98852_1002
xorg-randrproto=1.5.0=h7f98852_1001
xorg-recordproto=1.14.2=h7f98852_1002
xorg-renderproto=0.11.1=h7f98852_1002
xorg-util-macros=1.19.3=h7f98852_0
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xf86vidmodeproto=2.3.1=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.6=h166bdaf_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hfc55251_0
